itembank-api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
pgvector>=0.4.2

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Data processing
numpy>=2.0.0

# Development
tqdm>=4.67.0
